address Copilot review comments
- cugw_qkpt: throw if exactly one of {U, U_conj} is null in
  compute_second_tau_contraction (operand pair must be all-or-nothing);
  tighten U parameter docstring.
- cu_symmetry: rewrite the q_p0_transforms_conj comment so it points
  readers to the kernel's convention block rather than asserting a
  specific OP ↔ math correspondence at the storage site — the kernel
  actually uses OP_T on this buffer at GEMM 2c, which doesn't match the
  "OP_N → U_q†" reading the prior comment implied.
- cu_solver_test: cast result.size() to int for MPI_Bcast count
  (silent narrowing; safe in test, but flagged by -Wconversion).
- GW_X2C_Ar test data scripts: replace hard-coded \$HOME paths with
  GREEN_MBPT_PYTHON / GREEN_MBPT_BIN env vars (or local ScriptDir /
  GWBin overrides), with :? failure messages if neither is set.

+++ test/cu_solver_test.cpp
@@ -193,7 +193,7 @@ void check_x2c_ar_symmetry(const std::string& scf_type, const std::string& lin, 
       nt_out = 1;
     }
     // Broadcast result to all ranks so assertions run on consistent data.
-    MPI_Bcast(result.data(), result.size(), MPI_CXX_DOUBLE_COMPLEX, 0, green::utils::context().global);
+    MPI_Bcast(result.data(), static_cast<int>(result.size()), MPI_CXX_DOUBLE_COMPLEX, 0, green::utils::context().global);
     return std::make_pair(result, nt_out);
   };
